Improve handling of errors in ScalaCheck properties.
Before this commit, exceptions thrown in property evaluation would
only show the class and description, but no stack trace. This code
updates the `PropertyMatcher` classes to throw an `AsertionError`
containing the original exception and pretty description of the
property evaluation result if an exception is thrown in either the
property or the generator supplying it.